% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mak.R
\name{find_reactions}
\alias{find_reactions}
\title{Find Reactions in Stochiometric Matrices}
\usage{
find_reactions(makobj, i, sj)
}
\arguments{
\item{makobj}{Object of class \code{\link{mak}}.}

\item{i}{Index of target, single valued.}

\item{sj}{Set of signed parents.}
}
\value{
A logical vector indicating which rows of the stoichiometric matrices represent reactions satisfying target conditions.
}
\description{
Find reactions for which a given target species has parents among a certain set.
}
\details{
The target species is indicated via \code{i}. The signed parents \code{sj} must be indeces of the species, but a change of sign is allowed. Positive sign means positive interaction, while negative sign means negative interaction. The target  \code{i} has positive interaction with parent \code{j} in reaction \code{r} iff.  \eqn{B_{ri} - A_{ri} > 0} and \eqn{A_{rj} \neq 0}. Similarly, \code{i} has positive interaction with \code{j} in reaction \code{r} iff. \eqn{B_{ri} - A_{ri} < 0} and \eqn{A_{rj} \neq 0}. No interaction takes place if \eqn{A_{rj}=0}.

A reaction is accepted iff it either activates or inhibits \code{i}. A reaction \code{r} activates \code{i} iff. \code{i} interacts only with \code{i} and the positive elements of \code{sj} and at least one positive element of \code{sj} interacts with \code{i}. Similarly a reaction inhibits  \code{i} iff. \code{i} interacts only with \code{i} and the negative elements of \code{sj} and at least one negative element of \code{sj} interacts with \code{i}.
}
