% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfit.R
\name{gpfit}
\alias{gpfit}
\title{Gaussian Process Fits}
\usage{
gpfit(ys, tout)
}
\arguments{
\item{ys}{Data stored column-wise, first column must be time and each consequtive column is abundance of a species.}

\item{tout}{Time vector of desired timepoints at which to evaluate mean of gaussian process and its derivative.}
}
\value{
A list with two elements:
\item{ms}{Matrix of mean values, stored column wise.}
\item{dms}{Matrix of derivate mean values, stored columns wise.}
}
\description{
Returns mean estimates of gaussian process and its derivative
}
\details{
Whenever time (first column of \code{ys}) decreases, the system restarts. Hence the data is assumed generated from s different contexts, where s - 1 is the number of decreases in the time vector. Each context has its own initial condition and parameter vector. The number of decreases in time in \code{ys} must match the number of decreases in \code{tout}.
}
