% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mak.R
\name{join_reactions}
\alias{join_reactions}
\title{Join reactions of stoichiometric matrices}
\usage{
join_reactions(mak1, mak2)
}
\arguments{
\item{mak1}{\code{mak}-object.}

\item{mak2}{\code{mak}-object.}
}
\value{
An object of class \code{mak} holding the two stoichiometric matrices A and B.
}
\description{
rbinds the stoichiometric matrices of two \code{mak}-objects and removes duplets. Remaining variables are inherited from the first object.
}
