% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mak.R
\name{mak_enzyme}
\alias{mak_enzyme}
\title{Stoichiometric Matrices of Simple Enzymatic Reactions}
\usage{
mak_enzyme(d, empty = FALSE, ...)
}
\arguments{
\item{d}{Positive integer giving total number of species.}

\item{empty}{Logical indicating if empty elements on LHS and RHS are accepted.}

\item{...}{Other arguments passed to mak.}
}
\value{
An object of class \code{mak} holding the two stoichiometric matrices A and B.
}
\description{
Generate stoichiometric matrices of mass action kinetics reaction network containing all reactions on the form
\deqn{X + Y -> X + Z}
where \eqn{Y \neq Z \neq X} and \eqn{X, Y, Z \in 1,...,d} (\eqn{0,...,d} if \code{empty = TRUE})
}
