% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{parents}
\alias{parents}
\title{Generate all signed parent sets of certain size.}
\usage{
parents(d, i, np)
}
\arguments{
\item{d}{Positive integer giving total number of species.}

\item{i}{Positive integer, index of target.}

\item{np}{Positive integer, number of parents.}
}
\value{
A matrix, each column representing a unique signed parent set.
}
\description{
Generate all signed parent sets of certain size.
}
\examples{
parents(5, 2, 2)

library(magrittr)
p_max   <- 2
target  <- 1
m       <- mak_enzyme(5)
pas     <- sapply(seq_len(p_max), parents, d = ncol(m$A), i = target)
models  <- pas \%>\%
  lapply(. \%>\% apply(2, find_reactions, i = target, makobj = m)) \%>\%
  do.call(cbind, .)
# each column is one of the possible models where 'target' has between 1 and p_max parents
# each row represent a row (reaction) in A and B

# Non-magrittr version
models  <- do.call(cbind, lapply(pas, function(pa) apply(pa, 2, find_reactions, i = target, makobj = m)))
}
