% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmatch.R
\name{tsars}
\alias{tsars}
\title{Topological Sensitivity Analysis of MAK-Reaction Systems}
\usage{
tsars(makobj, y, max_reactions = 3, method = "exact", tout = NULL, ...)
}
\arguments{
\item{makobj}{Object of class \code{\link{mak}}.}

\item{y}{Matrix holding data. First column is time (must be increasing), remainder are species.}

\item{max_reactions}{Integer, maximal number of reactions, default is 5. Be very cautious, the number of models searched easily explodes if this is set too high.}

\item{method}{Character giving method for subset selection. Options: "exact" and "simanneal".}

\item{tout}{The (augmented) time-points used for gradient matching. If \code{NULL}, the observed time points are used.}
}
\value{
A list with an element for each species. Each of these elements is a list with the following elements
\itemize{
 \item{betas}{Parameter estimates for each possible model for said species in sparse matrix (row = model, col = rate parameter).}
 \item{rss}{Vector of residual sum of squares for each model.}
 \item{models}{Logical sparse matrix marking what rate parameters where included in each model (row = model, col = rate parameter). }
 \item{parents}{Matrix holding indeces of the parents considered in each model (row = model, col = parents). Signs mark activating/inhibiting role of that parent. 0-index marks lag of parent.}
}
}
\description{
Evaluates algorithm of Babtie et. al (2014, PNAS).
}
\details{
For each species all models based of the possible combinations of parents (at most \code{max_parents} per model) are found and fitted using gradient matching on a gaussian process fitted to data. Note the zero-model with no parents is trivial and thus not fitted.
}
